+++ src/main/java/com/smashingmods/chemlib/registry/ChemicalRegistry.java
@@ -76,7 +76,7 @@ public class ChemicalRegistry {
                                 BlockRegistry.BLOCKS.register(String.format("%s_lamp_block", elementName), () -> new LampBlock(ResourceLocation.fromNamespaceAndPath(ChemLib.MODID, elementName), ChemicalBlockType.LAMP, BlockRegistry.LAMP_BLOCKS, BlockRegistry.LAMP_PROPERTIES));
                                 BlockRegistry.getRegistryObjectByName(String.format("%s_lamp_block", elementName)).ifPresent(block -> ItemRegistry.fromChemicalBlock(block, new Item.Properties()));
                             }
-                            FluidRegistry.registerFluid(elementName, fluidTypePropertiesFactory(properties, ChemLib.MODID, elementName), Integer.parseInt(color, 16) | 0xFF000000, slopeFindDistance, decreasePerBlock);
+                            FluidRegistry.registerFluid(elementName, fluidTypePropertiesFactory(properties, ChemLib.MODID, elementName), Integer.parseInt(color, 16) | 0xFF000000, slopeFindDistance, decreasePerBlock, matterState == MatterState.GAS);
                         }
                     }
                 }
@@ -126,7 +126,7 @@ public class ChemicalRegistry {
                         int decreasePerBlock = properties.has("decrease_per_block") ? properties.get("decrease_per_block").getAsInt() : 1;
 
                         switch (matterState) {
-                            case LIQUID, GAS -> FluidRegistry.registerFluid(compoundName, fluidTypePropertiesFactory(properties, ChemLib.MODID, compoundName), Integer.parseInt(color, 16) | 0xFF000000, slopeFindDistance, decreasePerBlock);
+                            case LIQUID, GAS -> FluidRegistry.registerFluid(compoundName, fluidTypePropertiesFactory(properties, ChemLib.MODID, compoundName), Integer.parseInt(color, 16) | 0xFF000000, slopeFindDistance, decreasePerBlock, matterState == MatterState.GAS);
                         }
                     }
                 }