+++ changelog.md
@@ -3,3 +3,4 @@
 Changes:
 
 - Port to 1.21.1 neoforge
+- Gaseous fluids flow upwards